# contracts/doorstep/openapi.yaml
openapi: 3.0.3
info:
  title: Doorstep Home Services API
  version: 1.0.0
  description: >-
    Authoritative REST contract for Doorstep (Urban Company style home
    services): the customer surface (Momentum Android app and postbook-ui
    /doorstep), the professional surface (/pro, the doorstep_pro Android app)
    and the admin-internal family (/internal/admin, admin-service tokens only).


    Served by doorstep-service (port 8122, schema `doorstep`) behind the API
    gateway prefix /v1/doorstep (dormant gate DOORSTEP_PUBLIC_ENABLED, pilot
    allowlist DOORSTEP_PILOT_USER_IDS). Pilot city Hyderabad (`HYD`).


    Conventions (binding for every lane):

    - Every response is the shared envelope: success `{"data": <payload>,
      "meta": {"request_id": "..."}}`, failure `{"error": {"code", "message",
      "details"?}, "meta": {...}}`. Schemas below describe `data` unless they
      are named `*Envelope` or `ErrorEnvelope`.

    - Money is integer paise in fields ending `_paise`. Customer prices are
      GST-inclusive; the split is shown as `taxable_paise` + `tax_paise`.

    - Error codes are stable UPPER_SNAKE strings (see `x-doorstep-error-codes`).
      Clients branch on `error.code`, never on `message`.

    - Timestamps are RFC 3339 UTC. Dates are `YYYY-MM-DD` in the city's time
      zone (Asia/Kolkata). Times of day are `HH:MM`.

    - Optional values are emitted as explicit `null`, never omitted, on every
...

...
      gateway refuses /internal/ from the edge.


...


...
servers:
...
security:
  - gatewayIdentity: []
...
paths:
...
  /v1/doorstep/internal/admin/cities:
    get:
      summary: List cities
      operationId: adminListCities
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.read
      x-lane: A1
      responses:
        '200':
          description: Cities. Golden bytes admin_cities_list_200.json.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminCityList'}
    post:
      summary: Create a city
      operationId: adminCreateCity
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:config.write
      x-lane: A1
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminCityInput'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminCity'}
        '409':
          $ref: '#/components/responses/Error'
  /v1/doorstep/internal/admin/cities/{code}:
    patch:
      summary: Edit a city (settings, activate/deactivate)
      operationId: adminUpdateCity
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:config.write
      x-lane: A1
      parameters:
        - name: code
          in: path
          required: true
          schema: {type: string, example: HYD}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminCityPatch'}
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminCity'}
        '404':
          $ref: '#/components/responses/Error'
  /v1/doorstep/internal/admin/zones:
    get:
      summary: List zones (optionally by city)
      operationId: adminListZones
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.read
      x-lane: A1
      parameters:
        - $ref: '#/components/parameters/CityFilter'
      responses:
        '200':
          description: Zones
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminZoneList'}
    post:
      summary: Create a zone (GeoJSON Polygon or MultiPolygon, WGS84)
      operationId: adminCreateZone
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:config.write
      x-lane: A1
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminZoneInput'}
      responses:
        '201':
          description: Created. Golden bytes admin_zone_post_201.json.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminZone'}
        '422':
          description: DOORSTEP_ZONE_INVALID. Golden bytes admin_zone_post_422.json.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /v1/doorstep/internal/admin/zones/{id}:
    patch:
      summary: Edit a zone
      operationId: adminUpdateZone
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:config.write
      x-lane: A1
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminZonePatch'}
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminZone'}
  /v1/doorstep/internal/admin/categories:
    get:
      summary: List categories (all, incl. inactive)
      operationId: adminListCategories
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.read
      x-lane: A1
      responses:
        '200':
          description: Categories
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminCategoryList'}
    post:
      summary: Create a category
      operationId: adminCreateCategory
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.write
      x-lane: A1
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminCategoryInput'}
      responses:
        '201':
          description: Created. Golden bytes admin_category_post_201.json.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminCategory'}
        '409':
          description: DOORSTEP_CONFLICT (slug taken)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /v1/doorstep/internal/admin/categories/{id}:
    patch:
      summary: Edit a category
      operationId: adminUpdateCategory
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.write
      x-lane: A1
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminCategoryPatch'}
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminCategory'}
  /v1/doorstep/internal/admin/skills:
    get:
      summary: List skills
      operationId: adminListSkills
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.read
      x-lane: A1
      responses:
        '200':
          description: Skills
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: {$ref: '#/components/schemas/Skill'}
    post:
      summary: Create a skill
      operationId: adminCreateSkill
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.write
      x-lane: A1
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SkillInput'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Skill'}
  /v1/doorstep/internal/admin/services:
    get:
      summary: List services (optionally by category)
      operationId: adminListServices
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.read
      x-lane: A1
      parameters:
        - name: category_id
          in: query
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: Services
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminServiceList'}
    post:
      summary: Create a service
      operationId: adminCreateService
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.write
      x-lane: A1
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminServiceInput'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminService'}
  /v1/doorstep/internal/admin/services/{id}:
    get:
      summary: A service with all options, add-on groups and add-ons (incl. inactive)
      operationId: adminGetService
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.read
      x-lane: A1
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Service tree
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminServiceTree'}
    patch:
      summary: Edit a service
      operationId: adminUpdateService
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.write
      x-lane: A1
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminServicePatch'}
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminService'}
  /v1/doorstep/internal/admin/services/{id}/options:
    post:
      summary: Add an option to a service
      operationId: adminCreateOption
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.write
      x-lane: A1
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminOptionInput'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminOption'}
  /v1/doorstep/internal/admin/options/{id}:
    patch:
      summary: Edit an option
      operationId: adminUpdateOption
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.write
      x-lane: A1
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminOptionPatch'}
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminOption'}
  /v1/doorstep/internal/admin/services/{id}/addon-groups:
    post:
      summary: Add an add-on group to a service (min_select <= max_select, max_select >= 1)
      operationId: adminCreateAddonGroup
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.write
      x-lane: A1
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminAddonGroupInput'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminAddonGroup'}
  /v1/doorstep/internal/admin/addon-groups/{id}:
    patch:
      summary: Edit an add-on group
      operationId: adminUpdateAddonGroup
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.write
      x-lane: A1
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminAddonGroupPatch'}
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminAddonGroup'}
  /v1/doorstep/internal/admin/addon-groups/{id}/addons:
    post:
      summary: Add an add-on to a group
      operationId: adminCreateAddon
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.write
      x-lane: A1
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminAddonInput'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminAddon'}
  /v1/doorstep/internal/admin/addons/{id}:
    patch:
      summary: Edit an add-on
      operationId: adminUpdateAddon
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.write
      x-lane: A1
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminAddonPatch'}
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminAddon'}
  /v1/doorstep/internal/admin/prices:
    get:
      summary: Price rows (history included) for a city, optionally one item
      operationId: adminListPrices
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.read
      x-lane: A1
      parameters:
        - $ref: '#/components/parameters/CityFilter'
        - name: item_id
          in: query
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: Prices
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminPriceList'}
    post:
      summary: Set a price from effective_from (closes the open row of the same item and city)
      operationId: adminCreatePrice
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.write
      x-lane: A1
      description: >-
        Effective-dated. The currently open row (effective_to null) of the same
        (city, item) gets effective_to = new effective_from. A new row may not
        start before the open row's start; overlaps are refused by an
        exclusion constraint (DOORSTEP_PRICE_OVERLAP). Prices are GST-inclusive.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminPriceInput'}
      responses:
        '201':
          description: Created. Golden bytes admin_price_post_201.json.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminPrice'}
        '409':
          description: DOORSTEP_PRICE_OVERLAP
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /v1/doorstep/internal/admin/rate-cards:
    get:
      summary: Extras rate-card items (by city and/or category)
      operationId: adminListRateCards
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.read
      x-lane: A1
      parameters:
        - $ref: '#/components/parameters/CityFilter'
        - name: category_id
          in: query
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: Rate-card items
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminRateCardList'}
    post:
      summary: Create a rate-card item
      operationId: adminCreateRateCard
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.write
      x-lane: A1
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminRateCardInput'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminRateCard'}
  /v1/doorstep/internal/admin/rate-cards/{id}:
    patch:
      summary: Edit a rate-card item
      operationId: adminUpdateRateCard
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.write
      x-lane: A1
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminRateCardPatch'}
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminRateCard'}
  /v1/doorstep/internal/admin/slot-configs:
    get:
      summary: Slot configurations
      operationId: adminListSlotConfigs
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.read
      x-lane: A1
      parameters:
        - $ref: '#/components/parameters/CityFilter'
      responses:
        '200':
          description: Slot configs
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminSlotConfigList'}
    post:
      summary: Create a slot config (city default when category_id is null)
      operationId: adminCreateSlotConfig
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:config.write
      x-lane: A1
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminSlotConfigInput'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminSlotConfig'}
  /v1/doorstep/internal/admin/slot-configs/{id}:
    patch:
      summary: Edit a slot config
      operationId: adminUpdateSlotConfig
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:config.write
      x-lane: A1
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminSlotConfigPatch'}
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminSlotConfig'}
  /v1/doorstep/internal/admin/cancellation-rules:
    get:
      summary: Cancellation fee rules
      operationId: adminListCancellationRules
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.read
      x-lane: A1
      parameters:
        - $ref: '#/components/parameters/CityFilter'
      responses:
        '200':
          description: Rules
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminCancellationRuleList'}
    post:
      summary: Create a cancellation rule
      operationId: adminCreateCancellationRule
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:config.write
      x-lane: A1
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminCancellationRuleInput'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminCancellationRule'}
  /v1/doorstep/internal/admin/cancellation-rules/{id}:
    patch:
      summary: Edit a cancellation rule
      operationId: adminUpdateCancellationRule
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:config.write
      x-lane: A1
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminCancellationRulePatch'}
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminCancellationRule'}
  /v1/doorstep/internal/admin/commission-rules:
    get:
      summary: Commission rules
      operationId: adminListCommissionRules
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:catalogue.read
      x-lane: A1
      parameters:
        - $ref: '#/components/parameters/CityFilter'
      responses:
        '200':
          description: Rules
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminCommissionRuleList'}
    post:
      summary: Create a commission rule
      operationId: adminCreateCommissionRule
      tags: [Admin]
      security: [{adminServiceToken: []}]
      x-permission: doorstep:config.write
      x-lane: A1
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminCommissionRuleInput'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminCommissionRule'}
...
components:
  securitySchemes:
    gatewayIdentity:
      type: apiKey
      in: header
      name: X-Verified-User-Id
      description: Stamped by the gateway from the verified JWT (with X-Internal-Service-Key). Never sent by clients directly.
    adminServiceToken:
      type: apiKey
      in: header
      name: X-Service-Authorization
      description: 'Bearer <ed25519 service token>: iss admin-service, aud doorstep, scope [one doorstep:* permission], act <admin user id>, exp <= 60 s.'
...